// frontend/src/types/Entertainer.ts
export interface Entertainer {
    entertainerID: number;
    entStageName: string;
    entSSN: string | null;
    entStreetAddress: string | null;
    entCity: string | null;
    entState: string | null;
    entZipCode: string | null;
    entPhoneNumber: string | null;
    entWebPage: string | null;
    entEMailAddress: string | null;
    dateEntered: string | null;
  }
  